Implemented paycheck calculation.

+++ src/main/java/com/hrapplication/Payment.java
@@ -7,5 +7,5 @@ import com.hrapplication.Model.Employee;
  */
 public interface Payment {
 
-    int calcPay(Employee employee);
+    double calcPay (Employee employee);
 }
